automate it

+++ page2.php
@@ -2,18 +2,24 @@
 
 function readCsv($csvFile)
 {
-    $openFile = fopen($csvFile, "r");
-    while (($data = fgetcsv($openFile)) !== false) {
-        for ($i = 0; $i <= count($data); $i++) {
-            if ($i < count($data) - 1) {
-                print_r($data[$i]);
-                echo '; ';
-            } else {
-                print_r($data[$i]);
+    if (!is_file($csvFile)) {
+        header("Location: page1.php");
+        die();
+    } else {
+        $openFile = fopen($csvFile, "r");
+        while (($data = fgetcsv($openFile)) !== false) {
+            for ($i = 0; $i <= count($data); $i++) {
+                if ($i < count($data) - 1) {
+                    print_r($data[$i]);
+                    echo '; ';
+                } else {
+                    print_r($data[$i]);
+                }
             }
+            echo '</br>';
         }
-        echo '</br>';
     }
+
 }
 ;
